requirements.txt
gradio==4.44.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
joblib==1.3.2
pillow==10.2.0
tensorflow==2.15.0
transformers==4.39.3
